requirements.txt
absl-py==1.3.0
ale-py==0.7.5
attrs==22.1.0
box2d-py==2.3.5
cffi==1.15.1
cloudpickle==2.2.0
cycler==0.11.0
Cython==0.29.32
fasteners==0.18
fonttools==4.38.0
future==0.18.2
glfw==2.5.5
gym==0.21.0
gym-notices==0.0.8
gym-retro==0.8.0
imageio==2.22.4
importlib-metadata==4.13.0
importlib-resources==5.10.1
iniconfig==1.1.1
kiwisolver==1.4.4
lz4==4.0.2
matplotlib==3.5.3
mujoco==2.2.0
mujoco-py==2.1.2.14
numpy==1.18.0
opencv-python==4.6.0.66
packaging==22.0
Pillow==9.3.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pygame==2.1.0
pyglet==1.5.11
PyOpenGL==3.1.6
pyparsing==3.0.9
pytest==7.0.1
python-dateutil==2.8.2
six==1.16.0
swig==4.1.1
tomli==2.0.1
typing_extensions==4.4.0
zipp==3.11.0
